+++ src/Report/Bucket.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the pagespeed/api package.
+ *
+ * (c) Simon Andre <smn.andre@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace PageSpeed\Api\Report;
+
+/**
+ * Core Web Vitals rating bucket, mirroring the CrUX metric "category" strings.
+ */
+enum Bucket: string
+{
+    case Fast = 'FAST';
+    case Average = 'AVERAGE';
+    case Slow = 'SLOW';
+}